Updates the create command to support moving from @aha-app/aha-cli -&gt; aha-cli. The previous method of importing the ts definition file does not work with a non-namespaced package
